requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
attrs==24.2.0
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.62.3
hiredis==3.0.0
httplib2==0.22.0
idna==3.8
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
paramiko==3.4.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyNaCl==1.5.0
pyparsing==3.1.4
pypdfium2==4.30.0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.32
state==0.1.1.dev0
storage==0.0.4.3
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.7
